refactor: remove k8s & eks helm hooks & cleanup charts
