--- charts/eks/templates/pre-install-hook-job-rolebinding.yaml
@@ -1,24 +1,0 @@
-{{- if not .Values.headless }}
-{{- if .Values.job.enabled }}
-kind: RoleBinding
-apiVersion: rbac.authorization.k8s.io/v1
-metadata:
-  name: {{ .Release.Name }}-job
-  namespace: {{ .Release.Namespace }}
-  annotations:
-  {{- if .Values.globalAnnotations}}
-{{ toYaml .Values.globalAnnotations | indent 4 }}
-  {{- end }}
-    "helm.sh/hook": pre-install
-    "helm.sh/hook-weight": "3"
-    "helm.sh/hook-delete-policy": before-hook-creation,hook-succeeded
-subjects:
-  - kind: ServiceAccount
-    name: {{ .Release.Name }}-job
-    namespace: {{ .Release.Namespace }}
-roleRef:
-  kind: Role
-  name: {{ .Release.Name }}-job
-  apiGroup: rbac.authorization.k8s.io
-{{- end }}
-{{- end }}
