--- charts/eks/templates/pre-install-hook-job-serviceaccount.yaml
@@ -1,20 +1,0 @@
-{{- if not .Values.headless }}
-{{- if .Values.job.enabled }}
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: {{ .Release.Name }}-job
-  namespace: {{ .Release.Namespace }}
-  annotations:
-  {{- if .Values.globalAnnotations}}
-{{ toYaml .Values.globalAnnotations | indent 4 }}
-  {{- end }}
-    "helm.sh/hook": pre-install
-    "helm.sh/hook-weight": "3"
-    "helm.sh/hook-delete-policy": before-hook-creation,hook-succeeded
-{{- if .Values.serviceAccount.imagePullSecrets }}
-imagePullSecrets:
-{{ toYaml .Values.serviceAccount.imagePullSecrets | indent 2 }}
-{{- end }}
-{{- end }}
-{{- end }}
