+++ charts/eks/templates/syncer-deployment.yaml
@@ -75,8 +75,7 @@ spec:
         - name: tmp
           emptyDir: {}
         - name: certs
-          secret:
-            secretName: {{ .Release.Name }}-certs
+          emptyDir: {}
       {{- if .Values.syncer.volumes }}
 {{ toYaml .Values.syncer.volumes | indent 8 }}
       {{- end }}
@@ -234,7 +233,11 @@ spec:
             mountPath: /.cache/helm
           - name: tmp
             mountPath: /tmp
+          - mountPath: /pki
+            name: certs
+        {{- if .Values.syncer.volumeMounts }}
 {{ toYaml .Values.syncer.volumeMounts | indent 10 }}
+        {{- end }}
         {{- if .Values.syncer.extraVolumeMounts }}
 {{ toYaml .Values.syncer.extraVolumeMounts | indent 10 }}
         {{- end }}
