+++ charts/k0s/values.yaml
@@ -279,7 +279,6 @@ ingress:
   enabled: false
   # Ingress path type
   pathType: ImplementationSpecific
-  apiVersion: networking.k8s.io/v1
   ingressClassName: ""
   host: vcluster.local
   annotations:
