+++ charts/k3s/values.yaml
@@ -296,7 +296,6 @@ ingress:
   enabled: false
   # Ingress path type
   pathType: ImplementationSpecific
-  apiVersion: networking.k8s.io/v1
   ingressClassName: ""
   host: vcluster.local
   annotations:
