+++ charts/k8s/templates/ingress.yaml
@@ -1,5 +1,5 @@
 {{- if .Values.ingress.enabled }}
-apiVersion: {{ .Values.ingress.apiVersion }}
+apiVersion: networking.k8s.io/v1
 kind: Ingress
 metadata:
   {{- $annotations := merge .Values.ingress.annotations .Values.globalAnnotations }}
