--- charts/k8s/templates/pre-install-hook-job-role.yaml
@@ -1,20 +1,0 @@
-{{- if not .Values.headless }}
-{{- if .Values.job.enabled }}
-kind: Role
-apiVersion: rbac.authorization.k8s.io/v1
-metadata:
-  name: {{ .Release.Name }}-job
-  namespace: {{ .Release.Namespace }}
-  annotations:
-  {{- if .Values.globalAnnotations }}
-{{ toYaml .Values.globalAnnotations | indent 4 }}
-  {{- end }} 
-    "helm.sh/hook": pre-install
-    "helm.sh/hook-weight": "3"
-    "helm.sh/hook-delete-policy": before-hook-creation,hook-succeeded
-rules:
-  - apiGroups: [""]
-    resources: ["secrets", "configmaps","services"]
-    verbs: ["create", "get", "list"]
-{{- end }}
-{{- end }}
