--- charts/k8s/templates/pre-install-hook-job.yaml
@@ -1,86 +1,0 @@
-{{- if not .Values.headless }}
-{{- if .Values.job.enabled }}
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: {{ .Release.Name }}-job
-  namespace: {{ .Release.Namespace }}
-  annotations:
-  {{- if .Values.globalAnnotations }}
-{{ toYaml .Values.globalAnnotations | indent 4 }}
-  {{- end }}
-    "helm.sh/hook": pre-install
-    "helm.sh/hook-weight": "3"
-    "helm.sh/hook-delete-policy": before-hook-creation,hook-succeeded
-spec:
-  backoffLimit: 3
-  template:
-    metadata:
-      name: {{ .Release.Name }}-job
-      {{- if .Values.job.podAnnotations }}
-      annotations:
-{{ toYaml .Values.job.podAnnotations | indent 8 }}
-      {{- end }}
-      {{- if .Values.job.podLabels }}
-      labels:
-{{ toYaml .Values.job.podLabels | indent 8 }}
-      {{- end }}
-    spec:
-      serviceAccountName: {{ .Release.Name }}-job
-      restartPolicy: OnFailure
-      nodeSelector:
-{{ toYaml .Values.job.nodeSelector | indent 8 }}
-      affinity:
-{{ toYaml .Values.job.affinity | indent 8 }}
-      tolerations:
-{{ toYaml .Values.job.tolerations | indent 8 }}
-      {{- if .Values.job.priorityClassName }}
-      priorityClassName: {{ .Values.job.priorityClassName }}
-      {{- end }}
-      {{- if .Values.isolation.enabled }}
-      securityContext:
-        seccompProfile:
-          type: RuntimeDefault
-      {{- end }}
-      containers:
-        - name: certs
-          {{- if .Values.syncer.image }}
-          image: "{{ .Values.defaultImageRegistry }}{{ .Values.syncer.image }}"
-          {{- else }}
-          {{- if .Values.pro }}
-          image: "{{ .Values.defaultImageRegistry }}ghcr.io/loft-sh/vcluster-pro:{{ .Chart.Version }}"
-          {{- else }}
-          image: "{{ .Values.defaultImageRegistry }}ghcr.io/loft-sh/vcluster:{{ .Chart.Version }}"
-          {{- end }}
-          {{- end }}
-          imagePullPolicy: IfNotPresent
-          command:
-            - /vcluster
-            - certs
-          args:
-            - --prefix={{ .Release.Name }}
-            - --etcd-replicas={{ .Values.etcd.replicas }}
-            {{- if .Values.serviceCIDR }}
-            - --service-cidr={{ .Values.serviceCIDR }}
-            {{- end }}
-            {{- if .Values.job.extraArgs }}
-{{ toYaml .Values.job.extraArgs | indent 12 }}
-            {{- end }}
-          securityContext:
-{{ toYaml .Values.job.securityContext | indent 12 }}
-          volumeMounts:
-            - name: cert-storage
-              mountPath: /certs
-          {{- if .Values.job.volumeMounts }}
-{{ toYaml .Values.job.volumeMounts | indent 12 }}
-          {{- end }}
-          resources:
-{{ toYaml .Values.job.resources | indent 12 }}
-      volumes:
-        - name: cert-storage
-          emptyDir: {}
-      {{- if .Values.job.volumes }}
-{{ toYaml .Values.job.volumes | indent 8 }}
-      {{- end }}
-{{- end }}
-{{- end }}
