+++ charts/k8s/values.yaml
@@ -138,10 +138,7 @@ syncer:
   # image: ghcr.io/loft-sh/vcluster
   imagePullPolicy: ""
   extraArgs: []
-  volumeMounts:
-    - mountPath: /pki
-      name: certs
-      readOnly: true
+  volumeMounts: []
   extraVolumeMounts: []
   env: []
   livenessProbe:
@@ -344,37 +341,12 @@ service:
   # Set the loadBalancerClass if using an external load balancer controller
   loadBalancerClass: ""
 
-# job configuration
-job:
-  enabled: true
-  volumes: []
-  volumeMounts: []
-  extraArgs: []
-  priorityClassName: ""
-  nodeSelector: {}
-  affinity: {}
-  tolerations: []
-  resources: {}
-  podAnnotations: {}
-  podLabels: {}
-  # Security context configuration of the container
-  securityContext:
-    capabilities:
-      drop:
-        - ALL
-    allowPrivilegeEscalation: false
-    readOnlyRootFilesystem: true
-    runAsNonRoot: true
-    runAsUser: 10001
-    runAsGroup: 10001
-
 # Configure the ingress resource that allows you to access the vcluster
 ingress:
   # Enable ingress record generation
   enabled: false
   # Ingress path type
   pathType: ImplementationSpecific
-  apiVersion: networking.k8s.io/v1
   ingressClassName: ""
   host: vcluster.local
   annotations:
