+++ cmd/vcluster/cmd/root.go
@@ -21,6 +21,5 @@ func BuildRoot() *cobra.Command {
 
 	// add top level commands
 	rootCmd.AddCommand(NewStartCommand())
-	rootCmd.AddCommand(NewCertsCommand())
 	return rootCmd
 }
