+++ pkg/certs/kubeconfig.go
@@ -17,7 +17,6 @@ import (
 	"bytes"
 	"crypto"
 	"crypto/x509"
-	"fmt"
 	"os"
 	"path/filepath"
 	"time"
@@ -71,7 +70,7 @@ func CreateJoinControlPlaneKubeConfigFiles(outDir string, cfg *InitConfiguration
 
 	for _, file := range files {
 		if externaCA {
-			fmt.Printf("[kubeconfig] External CA mode: Using user provided %s\n", file)
+			klog.Infof("[kubeconfig] External CA mode: Using user provided %s", file)
 			continue
 		}
 		if err := createKubeConfigFiles(outDir, cfg, file); err != nil {
@@ -241,7 +240,7 @@ func createKubeConfigFileIfNotExists(outDir, filename string, config *clientcmda
 		if !os.IsNotExist(err) {
 			return err
 		}
-		fmt.Printf("[kubeconfig] Writing %q kubeconfig file\n", filename)
+		klog.Infof("[kubeconfig] Writing %q kubeconfig file", filename)
 		err = WriteToDisk(kubeConfigFilePath, config)
 		if err != nil {
 			return errors.Wrapf(err, "failed to save kubeconfig file %q on disk", kubeConfigFilePath)
@@ -251,7 +250,7 @@ func createKubeConfigFileIfNotExists(outDir, filename string, config *clientcmda
 	// kubeadm doesn't validate the existing kubeconfig file more than this (kubeadm trusts the client certs to be valid)
 	// Basically, if we find a kubeconfig file with the same path; the same CA cert and the same server URL;
 	// kubeadm thinks those files are equal and doesn't bother writing a new file
-	fmt.Printf("[kubeconfig] Using existing kubeconfig file: %q\n", kubeConfigFilePath)
+	klog.Infof("[kubeconfig] Using existing kubeconfig file: %q", kubeConfigFilePath)
 
 	return nil
 }
