+++ pkg/controllers/k8sdefaultendpoint/register.go
@@ -1,13 +1,13 @@
 package k8sdefaultendpoint
 
 import (
-	controllercontext "github.com/loft-sh/vcluster/cmd/vcluster/context"
+	"github.com/loft-sh/vcluster/pkg/setup/options"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/client-go/discovery"
 	"k8s.io/klog/v2"
 )
 
-func Register(ctx *controllercontext.ControllerContext) error {
+func Register(ctx *options.ControllerContext) error {
 	discoveryClient, err := discovery.NewDiscoveryClientForConfig(ctx.VirtualManager.GetConfig())
 	if err != nil {
 		return err
